Fix the error in the TDL_Object WriteText.
Reviewed-on: https://gerrit-ai.sophgo.vip:8443/148640

+++ include/c_apis/tdl_types.h
@@ -220,9 +220,9 @@ typedef struct {
 } TDLIspMeta;
 
 typedef struct {
-  float r;
-  float g;
-  float b;
+  int r;
+  int g;
+  int b;
 } color_rgb;
 
 typedef struct {
